+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
-name = "data-generators"
-version = "0.1.0"
+name = "vcf-generator"
+version = "0.2.0"
 description = ""
 authors = ["Bryan Wodi <12839540+talk2bryan@users.noreply.github.com>"]
 readme = "README.md"
@@ -9,7 +9,6 @@ readme = "README.md"
 python = "^3.10"
 fire = "^0.6.0"
 faker = "^24.4.0"
-
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^3.7.0"
