+++ vcf_generator/gen_contacts.py
@@ -4,11 +4,11 @@ of the Contact class with different field values."""
 import fire
 from faker import Faker
 
-from data_generators.contacts import MIN_NUM_CONTACTS
-from data_generators.contacts.address import Address
-from data_generators.contacts.contact import Contact
-from data_generators.contacts.email import Email
-from data_generators.contacts.phone_number import PhoneNumber
+from vcf_generator import MIN_NUM_CONTACTS
+from vcf_generator.address import Address
+from vcf_generator.contact import Contact
+from vcf_generator.email import Email
+from vcf_generator.phone_number import PhoneNumber
 
 fake = Faker()
